move tests to core
