+++ src/NServiceBus.AcceptanceTests/Core/Sagas/When_a_saga_is_completed.cs
@@ -1,32 +1,38 @@
-﻿namespace NServiceBus.AcceptanceTests.Sagas
+﻿namespace NServiceBus.AcceptanceTests.Core.Sagas
 {
     using System;
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
+    using Features;
     using NUnit.Framework;
 
-    public class When_completing_a_saga : NServiceBusAcceptanceTest
+    public class When_a_saga_is_completed : NServiceBusAcceptanceTest
     {
         [Test]
-        public async Task Should_hydrate_and_complete_the_existing_instance()
+        public async Task Saga_should_not_handle_subsequent_messages_for_that_sagadata()
         {
             var context = await Scenario.Define<Context>(c => { c.Id = Guid.NewGuid(); })
                 .WithEndpoint<ReceiveCompletesSagaEndpoint>(b =>
                 {
-                    b.When((session, ctx) => session.SendLocal(new StartSagaMessage
+                    b.When((session, c) => session.SendLocal(new StartSagaMessage
                     {
-                        SomeId = ctx.Id
+                        SomeId = c.Id
                     }));
-                    b.When(ctx => ctx.StartSagaMessageReceived, (session, c) => session.SendLocal(new CompleteSagaMessage
+                    b.When(c => c.StartSagaMessageReceived, (session, c) => session.SendLocal(new CompleteSagaMessage
+                    {
+                        SomeId = c.Id
+                    }));
+                    b.When(c => c.SagaCompleted, (session, c) => session.SendLocal(new AnotherMessage
                     {
                         SomeId = c.Id
                     }));
                 })
-                .Done(c => c.SagaCompleted)
+                .Done(c => c.AnotherMessageReceived)
                 .Run();
 
-            Assert.True(context.SagaCompleted);
+            Assert.True(context.AnotherMessageReceived, "AnotherMessage should have been delivered to the handler outside the saga");
+            Assert.False(context.SagaReceivedAnotherMessage, "AnotherMessage should not be delivered to the saga after completion");
         }
 
         public class Context : ScenarioContext
@@ -34,6 +40,8 @@
             public Guid Id { get; set; }
             public bool StartSagaMessageReceived { get; set; }
             public bool SagaCompleted { get; set; }
+            public bool AnotherMessageReceived { get; set; }
+            public bool SagaReceivedAnotherMessage { get; set; }
         }
 
         public class ReceiveCompletesSagaEndpoint : EndpointConfigurationBuilder
@@ -42,6 +50,7 @@
             {
                 EndpointSetup<DefaultServer>(b =>
                 {
+                    b.EnableFeature<TimeoutManager>();
                     b.ExecuteTheseHandlersFirst(typeof(TestSaga10));
                     b.LimitMessageProcessingConcurrencyTo(1); // This test only works if the endpoints processes messages sequentially
                 });
@@ -49,16 +58,21 @@
 
             public class TestSaga10 : Saga<TestSagaData10>,
                 IAmStartedByMessages<StartSagaMessage>,
-                IHandleMessages<CompleteSagaMessage>
+                IHandleMessages<CompleteSagaMessage>,
+                IHandleMessages<AnotherMessage>
             {
                 public Context Context { get; set; }
 
                 public Task Handle(StartSagaMessage message, IMessageHandlerContext context)
                 {
                     Data.SomeId = message.SomeId;
-
                     Context.StartSagaMessageReceived = true;
+                    return Task.FromResult(0);
+                }
 
+                public Task Handle(AnotherMessage message, IMessageHandlerContext context)
+                {
+                    Context.SagaReceivedAnotherMessage = true;
                     return Task.FromResult(0);
                 }
 
@@ -75,6 +89,8 @@
                         .ToSaga(s => s.SomeId);
                     mapper.ConfigureMapping<CompleteSagaMessage>(m => m.SomeId)
                         .ToSaga(s => s.SomeId);
+                    mapper.ConfigureMapping<AnotherMessage>(m => m.SomeId)
+                        .ToSaga(s => s.SomeId);
                 }
             }
 
@@ -84,12 +100,28 @@
             }
         }
 
+        public class CompletionHandler : IHandleMessages<AnotherMessage>
+        {
+            public Context Context { get; set; }
+
+            public Task Handle(AnotherMessage message, IMessageHandlerContext context)
+            {
+                Context.AnotherMessageReceived = true;
+                return Task.FromResult(0);
+            }
+        }
+
         public class StartSagaMessage : ICommand
         {
             public Guid SomeId { get; set; }
         }
 
         public class CompleteSagaMessage : ICommand
+        {
+            public Guid SomeId { get; set; }
+        }
+
+        public class AnotherMessage : ICommand
         {
             public Guid SomeId { get; set; }
         }
