+++ src/NServiceBus.AcceptanceTests/Core/Sagas/When_completing_a_saga.cs
@@ -1,38 +1,32 @@
-﻿namespace NServiceBus.AcceptanceTests.Sagas
+﻿namespace NServiceBus.AcceptanceTests.Core.Sagas
 {
     using System;
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
-    using Features;
     using NUnit.Framework;
 
-    public class When_a_saga_is_completed : NServiceBusAcceptanceTest
+    public class When_completing_a_saga : NServiceBusAcceptanceTest
     {
         [Test]
-        public async Task Saga_should_not_handle_subsequent_messages_for_that_sagadata()
+        public async Task Should_hydrate_and_complete_the_existing_instance()
         {
             var context = await Scenario.Define<Context>(c => { c.Id = Guid.NewGuid(); })
                 .WithEndpoint<ReceiveCompletesSagaEndpoint>(b =>
                 {
-                    b.When((session, c) => session.SendLocal(new StartSagaMessage
+                    b.When((session, ctx) => session.SendLocal(new StartSagaMessage
                     {
-                        SomeId = c.Id
+                        SomeId = ctx.Id
                     }));
-                    b.When(c => c.StartSagaMessageReceived, (session, c) => session.SendLocal(new CompleteSagaMessage
-                    {
-                        SomeId = c.Id
-                    }));
-                    b.When(c => c.SagaCompleted, (session, c) => session.SendLocal(new AnotherMessage
+                    b.When(ctx => ctx.StartSagaMessageReceived, (session, c) => session.SendLocal(new CompleteSagaMessage
                     {
                         SomeId = c.Id
                     }));
                 })
-                .Done(c => c.AnotherMessageReceived)
+                .Done(c => c.SagaCompleted)
                 .Run();
 
-            Assert.True(context.AnotherMessageReceived, "AnotherMessage should have been delivered to the handler outside the saga");
-            Assert.False(context.SagaReceivedAnotherMessage, "AnotherMessage should not be delivered to the saga after completion");
+            Assert.True(context.SagaCompleted);
         }
 
         public class Context : ScenarioContext
@@ -40,8 +34,6 @@
             public Guid Id { get; set; }
             public bool StartSagaMessageReceived { get; set; }
             public bool SagaCompleted { get; set; }
-            public bool AnotherMessageReceived { get; set; }
-            public bool SagaReceivedAnotherMessage { get; set; }
         }
 
         public class ReceiveCompletesSagaEndpoint : EndpointConfigurationBuilder
@@ -50,7 +42,6 @@
             {
                 EndpointSetup<DefaultServer>(b =>
                 {
-                    b.EnableFeature<TimeoutManager>();
                     b.ExecuteTheseHandlersFirst(typeof(TestSaga10));
                     b.LimitMessageProcessingConcurrencyTo(1); // This test only works if the endpoints processes messages sequentially
                 });
@@ -58,21 +49,16 @@
 
             public class TestSaga10 : Saga<TestSagaData10>,
                 IAmStartedByMessages<StartSagaMessage>,
-                IHandleMessages<CompleteSagaMessage>,
-                IHandleMessages<AnotherMessage>
+                IHandleMessages<CompleteSagaMessage>
             {
                 public Context Context { get; set; }
 
                 public Task Handle(StartSagaMessage message, IMessageHandlerContext context)
                 {
                     Data.SomeId = message.SomeId;
-                    Context.StartSagaMessageReceived = true;
-                    return Task.FromResult(0);
-                }
 
-                public Task Handle(AnotherMessage message, IMessageHandlerContext context)
-                {
-                    Context.SagaReceivedAnotherMessage = true;
+                    Context.StartSagaMessageReceived = true;
+
                     return Task.FromResult(0);
                 }
 
@@ -89,8 +75,6 @@
                         .ToSaga(s => s.SomeId);
                     mapper.ConfigureMapping<CompleteSagaMessage>(m => m.SomeId)
                         .ToSaga(s => s.SomeId);
-                    mapper.ConfigureMapping<AnotherMessage>(m => m.SomeId)
-                        .ToSaga(s => s.SomeId);
                 }
             }
 
@@ -100,28 +84,12 @@
             }
         }
 
-        public class CompletionHandler : IHandleMessages<AnotherMessage>
-        {
-            public Context Context { get; set; }
-
-            public Task Handle(AnotherMessage message, IMessageHandlerContext context)
-            {
-                Context.AnotherMessageReceived = true;
-                return Task.FromResult(0);
-            }
-        }
-
         public class StartSagaMessage : ICommand
         {
             public Guid SomeId { get; set; }
         }
 
         public class CompleteSagaMessage : ICommand
-        {
-            public Guid SomeId { get; set; }
-        }
-
-        public class AnotherMessage : ICommand
         {
             public Guid SomeId { get; set; }
         }
